feat(peso): venta y stock por kg en POS e inventario
Cambios principales:

Schema (Prisma)
- Product.stock/minStock/maxStock: Int → Decimal(12,3)
- SaleItem.quantity, StockMovement.{quantity,stockBefore,stockAfter},
  ProductLoss.quantity: Int → Decimal(12,3)
- Combos y recargas siguen siendo Int (siempre por unidad)
- db push con --accept-data-loss hace el cast automático

POS
- Modal nuevo (WeightInputModal): cuando agregás un producto sold-by-weight,
  pide el peso en kg con quick buttons (250g/500g/1kg/2kg) y muestra
  total en vivo. Acepta hasta 3 decimales (gramo)
- POS detecta soldByWeight tanto en click directo como en escaneo y abre
  el modal en lugar de meter quantity=1
- CartPanel muestra peso editable (botón "0.350 kg") en lugar del +/-
  para items por peso. Click → reabre modal para re-pesar
- PaymentModal: para MP arma 1 item con subtotal completo (MP exige int);
  para venta interna manda quantity decimal

Inventario
- Display de stock: "5.000 kg" para por-peso, integer para unidades
- ProductModal acepta decimales en stock/minStock cuando soldByWeight=true
  (step=0.001) y label aclara "(kg)"

Plumbing
- lib/db.ts: extensión de Prisma Client que coerciona Decimal → number en
  resultados (stock, quantity, etc.). Evita parche manual de Number() en
  todas las comparaciones server/client
- Zod schemas: quantity y stock pasan de int() a number() libre

+++ app/api/productos/bulk-create/route.ts
@@ -29,8 +29,8 @@ const inputSchema = z.object({
         barcode: z.string().nullable().optional(),
         salePrice: z.number().min(0).optional().default(0),
         costPrice: z.number().min(0).optional().default(0),
-        stock: z.number().int().min(0).optional().default(0),
-        minStock: z.number().int().min(0).optional().default(5),
+        stock: z.number().min(0).optional().default(0),
+        minStock: z.number().min(0).optional().default(5),
         categoryId: z.string().nullable().optional(),
         supplierId: z.string().nullable().optional(),
       })
